Implement image verification
Function checks if the file being uploaded is a valid file type (being
PNG, JPG, JPEG) - if so endpoint will raise 415 Unsupported Media Type.
Function will also check if the file being uploaded is bigger than
2.7MB - if so endpoint will raise 413 Entity too large.

If the file passes above checks the file is checked against maximum
dimensions - 1024x1024 - if so file will be transformed to a thumbnail.